Aspect ratio

+++ RayTracer/src/Render.cpp
@@ -21,7 +21,7 @@ void Renderer::OnResize(uint32_t width, uint32_t height)
 
 void Renderer::Render()
 {
-	float aspectRatio = m_FinalImage->GetWidth() / m_FinalImage->GetHeight();
+	float aspectRatio = (float)m_FinalImage->GetWidth() / (float)m_FinalImage->GetHeight();
 	for (uint32_t y = 0; y < m_FinalImage->GetHeight(); y++)
 	{
 		for (uint32_t x = 0; x < m_FinalImage->GetWidth(); x++)
